moves some functions to system info
